+++ noodles-csi/src/io/indexed_reader.rs
@@ -1,7 +1,4 @@
-use std::{
-    io::{self, Read, Seek},
-    str,
-};
+use std::io::{self, Read, Seek};
 
 use noodles_bgzf as bgzf;
 use noodles_core::Region;
@@ -64,12 +61,9 @@ where
             .header()
             .ok_or_else(|| io::Error::new(io::ErrorKind::InvalidInput, "missing index header"))?;
 
-        let region_name = str::from_utf8(region.name())
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
-
         let reference_sequence_id = header
             .reference_sequence_names()
-            .get_index_of(region_name)
+            .get_index_of(region.name())
             .ok_or_else(|| {
                 io::Error::new(
                     io::ErrorKind::InvalidInput,