+++ noodles-csi/src/io/writer/index/header.rs
@@ -69,6 +69,8 @@ where
 
 #[cfg(test)]
 mod tests {
+    use bstr::BString;
+
     use super::*;
 
     #[test]
@@ -78,7 +80,7 @@ mod tests {
         let expected = [0x00, 0x00, 0x00, 0x00];
         assert_eq!(buf, expected);
 
-        let names = [String::from("sq0"), String::from("sq1")]
+        let names = [BString::from("sq0"), BString::from("sq1")]
             .into_iter()
             .collect();
         let header = crate::binning_index::index::header::Builder::vcf()